Improves expiring ticket records report
Refactors the expiring tickets report to include filtering by legacy tickets and inactive employees.

This change enhances the report's flexibility and data accuracy. It allows users to include or exclude legacy tickets and inactive employees, providing a more comprehensive view of expiring tickets.

Adds toggles for legacy tickets and inactive employees.
Updates the ticket fetching and record fetching logic to incorporate the new filters.
Improves the user interface for a better user experience.
